fix: card delete

+++ controllers/cards.js
@@ -1,8 +1,9 @@
 const Card = require('../models/card');
 
 const { CREATED_CODE } = require('../utils/constants');
-const { BadRequest } = require('../errors/BadRequest');
-const { NotFoundError } = require('../errors/NotFoundError');
+const BadRequestError = require('../errors/BadRequestError');
+const NotFoundError = require('../errors/NotFoundError');
+const ForbiddenError = require('../errors/ForbiddenError');
 // const { ServerError } = require('../errors/ServerError');
 
 // GET /cards
@@ -26,7 +27,7 @@ const createCard = (req, res, next) => {
     .catch((err) => {
       if (err.name === 'ValidationError') {
         // res.status(BAD_REQUEST_CODE).send(validationErrorMessage);
-        throw next(new BadRequest('Ошибка валидации'));
+        throw next(new BadRequestError('Ошибка валидации'));
       // } else {
       //   // res.status(SERVER_ERROR_CODE).send(serverErrorMessage);
       //   throw next(new ServerError('Ошибка на стороне сервера'));
@@ -37,24 +38,31 @@ const createCard = (req, res, next) => {
 
 // DELETE /cards/:cardId
 const deleteCard = (req, res, next) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .then((card) => {
-      if (card == null) {
-        // res.status(NOT_FOUND_CODE).send(cardNotFoundMessage);
-        throw new NotFoundError('Карточка не найдена');
-      }
-      res.send(card);
-    })
+  console.log(req.params.cardId);
+  // console.log('req.params.cardId:', req.params.cardId, 'req.user._id:', req.user._id);
+  Card.findById(req.params.cardId)
+    .then(
+      (card) => {
+        console.log('CARD:', card);
+        if (!card) {
+          throw new NotFoundError('Карточка не найдена');
+        }
+
+        if (req.params.cardId.owner !== req.user._id) {
+          throw new ForbiddenError('Недостаточно прав');
+        }
+        // res.send(card);
+        Card.findByIdAndRemove(req.params.cardId)
+          .then((deletedCard) => res.send(deletedCard))
+          .catch(next);
+      },
+    )
     .catch((err) => {
       if (err.name === 'CastError') {
-        // res.status(BAD_REQUEST_CODE).send(incorrectCardIdMessage);
-        throw next(new BadRequest('Некорректный id карточки'));
-      // } else {
-      //   // res.status(SERVER_ERROR_CODE).send(serverErrorMessage);
-      //   throw next(new ServerError('Ошибка на стороне сервера'));
-      // }
-      } else throw next(err);
+        throw next(new BadRequestError('Некорректный id карточки'));
+      } else next(err);
     });
+  // .catch(next);
 };
 
 // PUT /cards/:cardId/likes
@@ -65,21 +73,22 @@ const putLike = (req, res, next) => {
     { new: true },
   )
     .then((card) => {
-      if (card == null) {
+      if (!card) {
         // res.status(NOT_FOUND_CODE).send(cardNotFoundMessage);
         throw new NotFoundError('Карточка не найдена');
       }
       res.send(card);
     })
     .catch((err) => {
+      console.log(err);
       if (err.name === 'CastError') {
         // res.status(BAD_REQUEST_CODE).send(incorrectCardIdMessage);
-        throw next(new BadRequest('Некорректный id карточки'));
+        throw next(new BadRequestError('Некорректный id карточки'));
       // } else {
       //   // res.status(SERVER_ERROR_CODE).send(serverErrorMessage);
       //   throw next(new ServerError('Ошибка на стороне сервера'));
       // }
-      } else throw next(err);
+      } else next(err);
     });
 };
 
@@ -91,7 +100,7 @@ const deleteLike = (req, res, next) => {
     { new: true },
   )
     .then((card) => {
-      if (card == null) {
+      if (!card) {
         // res.status(NOT_FOUND_CODE).send(cardNotFoundMessage);
         throw new NotFoundError('Карточка не найдена');
       }
@@ -100,12 +109,12 @@ const deleteLike = (req, res, next) => {
     .catch((err) => {
       if (err.name === 'CastError') {
         // res.status(BAD_REQUEST_CODE).send(incorrectCardIdMessage);
-        throw next(new BadRequest('Некорректный id карточки'));
+        throw next(new BadRequestError('Некорректный id карточки'));
       // } else {
       //   // res.status(SERVER_ERROR_CODE).send(serverErrorMessage);
       //   throw new ServerError('Ошибка на стороне сервера');
       // }
-      } else throw next(err);
+      } else next(err);
     });
 };
